requirements.txt
# Core dependencies
google-generativeai>=0.3.0
chromadb>=0.4.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20

# Data processing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0

# Text processing
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0

# CLI and utilities
click>=8.1.0
tqdm>=4.65.0
rich>=13.0.0

# Optional: For Jupyter demo
jupyter>=1.0.0
ipykernel>=6.25.0
